requirements.txt
python-telegram-bot==20.7
yfinance==0.2.54
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
python-dotenv==1.0.0
yookassa==3.2.1
aiohttp
cryptography
pyjwt
telegram
flask
websockets
